refactor(city): extract shared City/Map types to app/utils
Relocate types, bucketCityResults, EventEmitter, and exhibitionPeriodParser
out of Scenes/Map and into app/utils/cityGuide (and app/utils directly),
since they are shared across the City and Map scenes and violate the
project's cross-scene import convention.

Also removes now-dead @ts-expect-error suppressions in cityTabs.ts that
the newly explicit BucketResults/MapTab typing exposed as unused, and
relocates the Event and TabFairItemRow test files to __tests__/ (pulled
forward from the test-dedupe item since their fixtures reference the
relocated types module and would otherwise fail to compile once
Scenes/Map/types.ts is deleted).
