+++ .secrets.baseline
@@ -327,22 +327,13 @@
         "line_number": 81
       }
     ],
-    "src/app/Scenes/City/Components/Event/Event.tests.tsx": [
+    "src/app/Scenes/City/Components/Event/__tests__/Event.tests.tsx": [
       {
         "type": "Base64 High Entropy String",
-        "filename": "src/app/Scenes/City/Components/Event/Event.tests.tsx",
+        "filename": "src/app/Scenes/City/Components/Event/__tests__/Event.tests.tsx",
         "hashed_secret": "4d3f5429714b267f0714ed59737e6c2475826176",
         "is_verified": false,
-        "line_number": 7
-      }
-    ],
-    "src/app/Scenes/City/Components/Event/__tests__/index.tests.tsx": [
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "src/app/Scenes/City/Components/Event/__tests__/index.tests.tsx",
-        "hashed_secret": "4d3f5429714b267f0714ed59737e6c2475826176",
-        "is_verified": false,
-        "line_number": 7
+        "line_number": 8
       }
     ],
     "src/app/Scenes/City/Components/EventSection/__tests__/EventSection.tests.tsx": [
@@ -363,22 +354,13 @@
         "line_number": 8
       }
     ],
-    "src/app/Scenes/City/Components/TabFairItemRow/TabFairItemRow.tests.tsx": [
+    "src/app/Scenes/City/Components/TabFairItemRow/__tests__/TabFairItemRow.tests.tsx": [
       {
         "type": "Base64 High Entropy String",
-        "filename": "src/app/Scenes/City/Components/TabFairItemRow/TabFairItemRow.tests.tsx",
+        "filename": "src/app/Scenes/City/Components/TabFairItemRow/__tests__/TabFairItemRow.tests.tsx",
         "hashed_secret": "5d4d720117d9b923461eda104bdee573deb27d4e",
         "is_verified": false,
-        "line_number": 12
-      }
-    ],
-    "src/app/Scenes/City/Components/TabFairItemRow/__tests__/index.tests.tsx": [
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "src/app/Scenes/City/Components/TabFairItemRow/__tests__/index.tests.tsx",
-        "hashed_secret": "5d4d720117d9b923461eda104bdee573deb27d4e",
-        "is_verified": false,
-        "line_number": 12
+        "line_number": 13
       }
     ],
     "src/app/Scenes/Collection/Components/__fixtures__/CollectionFixture.ts": [
