+++ src/app/Scenes/City/City.tsx
@@ -1,8 +1,8 @@
 import { Flex, Tabs } from "@artsy/palette-mobile"
 import { TabsContainer } from "@artsy/palette-mobile/dist/elements/Tabs/TabsContainer"
-import { EventEmitter } from "app/Scenes/Map/EventEmitter"
-import { BucketKey, BucketResults } from "app/Scenes/Map/bucketCityResults"
-import { MapTab } from "app/Scenes/Map/types"
+import { EventEmitter } from "app/utils/cityGuide/EventEmitter"
+import { BucketKey, BucketResults } from "app/utils/cityGuide/bucketCityResults"
+import { MapTab } from "app/utils/cityGuide/types"
 import { Schema } from "app/utils/track"
 import React, { useEffect, useState } from "react"
 import { RelayProp } from "react-relay"
