+++ src/app/Scenes/City/CitySectionList.tsx
@@ -8,9 +8,9 @@ import {
 import { CitySectionList_viewer$key } from "__generated__/CitySectionList_viewer.graphql"
 import { LoadFailureView } from "app/Components/LoadFailureView"
 import { PAGE_SIZE } from "app/Components/constants"
-import { BucketKey } from "app/Scenes/Map/bucketCityResults"
-import { Show } from "app/Scenes/Map/types"
 import { goBack } from "app/system/navigation/navigate"
+import { BucketKey } from "app/utils/cityGuide/bucketCityResults"
+import { Show } from "app/utils/cityGuide/types"
 import { extractNodes } from "app/utils/extractNodes"
 import { withSuspense } from "app/utils/hooks/withSuspense"
 import { isCloseToBottom } from "app/utils/isCloseToBottom"
