+++ src/app/Scenes/City/Components/AllEvents.tsx
@@ -1,7 +1,7 @@
 import { Box, Separator, Spacer, Tabs, Text } from "@artsy/palette-mobile"
 import { BottomSheetScrollView } from "@gorhom/bottom-sheet"
 import { EventSection } from "app/Scenes/City/Components/EventSection/EventSection"
-import { BucketResults } from "app/Scenes/Map/bucketCityResults"
+import { BucketResults } from "app/utils/cityGuide/bucketCityResults"
 import { isEqual } from "lodash"
 import React, { Fragment } from "react"
 import { Platform, ViewProps } from "react-native"
