+++ src/app/Scenes/Partner/Components/PartnerShowRailItem.tsx
@@ -2,8 +2,8 @@ import { Flex, Spacer, Text, useScreenDimensions } from "@artsy/palette-mobile"
 import { PartnerShowRailItem_show$data } from "__generated__/PartnerShowRailItem_show.graphql"
 import { ImageWithFallback } from "app/Components/ImageWithFallback/ImageWithFallback"
 import { ShowFollowButton } from "app/Components/ShowFollowButton"
-import { exhibitionDates } from "app/Scenes/Map/exhibitionPeriodParser"
 import { RouterLink } from "app/system/navigation/RouterLink"
+import { exhibitionDates } from "app/utils/exhibitionPeriodParser"
 import { Schema } from "app/utils/track"
 import { first } from "lodash"
 import React from "react"
